6.best_practices/requirements.txt
boto3
mlflow
scikit-learn==1.5.0
pytest
deepdiff
pylint
black
isort
pre-commit
pandas
pickle
numpy==1.26.4
s3fs
fastparquet
awscli
python-dotenv
